requirements.txt
matplotlib==3.7.0
numpy==1.23.5
pandas==2.0.3
Pillow==10.0.1
plotly==5.13.1
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.11.2
streamlit==1.27.2
streamlit_option_menu==0.3.6